requirements.txt
cassandra-driver==3.29.1
click==8.1.7
geomet==0.2.1.post1
six==1.16.0
customtkinter==5.2.2
darkdetect==0.8.0
geomet==0.2.1.post1
iniconfig==2.0.0
jsonpointer==2.1
pluggy==1.5.0
pillow==10.3.0
setuptools==68.2.2
wheel==0.41.2
pytest==8.2.1